requirements.txt
attrs==16.0.0
cffi==1.6.0
cryptography==1.4
cssselect==0.9.2
decorator==4.0.10
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.7
django-dynamic-scraper==0.11.0
django-toolbelt==0.0.1
enum34==1.1.6
future==0.15.2
gunicorn==19.6.0
idna==2.1
ipaddress==1.0.16
jsonpath-rw==1.4.0
lxml==3.6.0
parsel==1.0.2
Pillow==3.2.0
ply==3.8
psycopg2==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
PyDispatcher==2.0.5
pyOpenSSL==16.0.0
queuelib==1.4.2
Scrapy==1.1.0
scrapy-djangoitem==1.1.1
service-identity==16.0.0
six==1.10.0
static3==0.7.0
Twisted==16.2.0
w3lib==1.14.2
whitenoise==3.2
zope.interface==4.2.0
